refactor: moved all of the logic from controllers into services, added transactions

+++ app/src/main/java/i/love/building/webapps/its/actually/my/favourite/thing/in/the/world/especially/in/java/service/UserService.java
@@ -4,14 +4,22 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import i.love.building.webapps.its.actually.my.favourite.thing.in.the.world.especially.in.java.common.exception.AlreadyExistsException;
+import i.love.building.webapps.its.actually.my.favourite.thing.in.the.world.especially.in.java.common.exception.ObjectNotFoundException;
+import i.love.building.webapps.its.actually.my.favourite.thing.in.the.world.especially.in.java.common.exception.user.CannotDeleteAdminException;
 import i.love.building.webapps.its.actually.my.favourite.thing.in.the.world.especially.in.java.model.User;
 import i.love.building.webapps.its.actually.my.favourite.thing.in.the.world.especially.in.java.repository.UserRepository;
+import jakarta.transaction.Transactional;
 
 @Service
 public class UserService {
+    @Value("#{environment.ADMIN_USERNAME}")
+    private String adminUsername;
+    
     @Autowired
     private UserRepository users;
 
@@ -27,7 +35,12 @@ public class UserService {
         return this.users.findByName(name);
     }
 
-    public User create(User user) {
+    @Transactional
+    public User create(User user) throws AlreadyExistsException {
+        Optional<User> existing = this.users.findByName(user.getName());
+        if (existing.isPresent()) {
+            throw new AlreadyExistsException("user", existing.get().getName());
+        }
         return this.users.save(user);
     }
 
@@ -36,7 +49,13 @@ public class UserService {
         return updated > 0;
     }
 
-    public boolean deleteById(Long id) {
+    @Transactional
+    public boolean deleteById(Long id) throws CannotDeleteAdminException, ObjectNotFoundException {
+        boolean isAdmin = this.users.findById(id).map(u -> u.getName() == this.adminUsername)
+            .orElseThrow(() -> new ObjectNotFoundException("user with id '%d'", id));
+        if (isAdmin) {
+            throw new CannotDeleteAdminException();
+        }
         int updated = this.users.deleteByIdReturning(id);
         return updated > 0;
     }
